fixes for issues discovered in static code analysis
