+++ compute.py
@@ -53,15 +53,6 @@ def sum_to_scale(values,
     return np.hstack(([np.NaN]*(scale - 1), sliding_sums))
 
 #-----------------------------------------------------------------------------------------------------------------------
-def _count_zeros_and_non_missings(values):
-    
-    # count the number of zeros and non-missing (non-NaN) values
-    zeros = values.size - np.count_nonzero(values)
-    non_missings = np.count_nonzero(~np.isnan(values))
-
-    return zeros, non_missings
-
-#-----------------------------------------------------------------------------------------------------------------------
 def _estimate_pearson3_parameters(lmoments):    
     '''
     Estimate parameters via L-moments for the Pearson Type III distribution, based on Fortran code written 
@@ -239,7 +230,7 @@ def _pearson3_fitting_values(values,
         calibration_values = values[calibration_begin_index:calibration_end_index, month_index]
 
         # count the number of zeros and valid (non-missing/non-NaN) values
-        number_of_zeros, number_of_non_missing = _count_zeros_and_non_missings(calibration_values)
+        number_of_zeros, number_of_non_missing = utils.count_zeros_and_non_missings(calibration_values)
 
         # make sure we have at least four values that are both non-missing (i.e. non-NaN)
         # and non-zero, otherwise use the entire period of record
@@ -249,7 +240,7 @@ def _pearson3_fitting_values(values,
             calibration_values = values[:, month_index]
             
             # get new counts of the zeros and non-missing values
-            number_of_zeros, number_of_non_missing = _count_zeros_and_non_missings(calibration_values)
+            number_of_zeros, number_of_non_missing = utils.count_zeros_and_non_missings(calibration_values)
             
         # calculate the probability of zero for the calendar month
         probability_of_zero = 0.0
@@ -685,7 +676,7 @@ def _pearson3_fitting_values_new(values,
         calibration_values = values[calibration_begin_index:calibration_end_index, month_index]
 
         # count the number of zeros and valid (non-missing/non-NaN) values
-        number_of_zeros, number_of_non_missing = _count_zeros_and_non_missings(calibration_values)
+        number_of_zeros, number_of_non_missing = utils.count_zeros_and_non_missings(calibration_values)
 
         # make sure we have at least four values that are both non-missing (i.e. non-NaN)
         # and non-zero, otherwise use the entire period of record
@@ -695,7 +686,7 @@ def _pearson3_fitting_values_new(values,
             calibration_values = values[:, month_index]
             
             # get new counts of the zeros and non-missing values
-            number_of_zeros, number_of_non_missing = _count_zeros_and_non_missings(calibration_values)
+            number_of_zeros, number_of_non_missing = utils.count_zeros_and_non_missings(calibration_values)
             
         # calculate the probability of zero for the calendar month
         probability_of_zero = 0.0
