+++ test_thornthwaite.py
@@ -289,3 +289,4 @@ class ThornthwaiteTestCase(unittest.TestCase):
 if __name__ == '__main__':
     
     unittest.main()
+
