+++ test_utils.py
@@ -1,7 +1,5 @@
 import logging
-import math
 import numpy as np
-import scipy
 import unittest
 import utils
 
@@ -43,7 +41,7 @@ class UtilsTestCase(unittest.TestCase):
         # exercise the function
         values_2d_computed = utils.reshape_to_years_months(values_2d)
         
-         # verify that the function performed as expected
+        # verify that the function performed as expected
         np.testing.assert_equal(values_2d_computed, 
                                 values_2d, 
                                 'Not returning a valid 2-D array as expected')
@@ -56,6 +54,17 @@ class UtilsTestCase(unittest.TestCase):
         # make sure that the function croaks with a ValueError
         np.testing.assert_raises(ValueError, utils.reshape_to_years_months, values_2d)
         
+    #----------------------------------------------------------------------------------------
+    def test_count_zeros_and_non_missings(self):
+        '''
+        Test for the utils.count_zeros_and_non_missings() function
+        '''
+        values = np.array([3, 4, 0, 2, 3.1, 5, np.NaN, 8, 5, 6, 0.0, np.NaN, 5.6, 2])
+        zeros, non_missings = utils.count_zeros_and_non_missings(values)
+        self.assertEqual(zeros, 2, 'Failed to correctly count zero values')
+        self.assertEqual(non_missings, 12, 'Failed to correctly count non-missing values')
+                
 #--------------------------------------------------------------------------------------------
 if __name__ == '__main__':
     unittest.main()
+
